backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
